drm/panel: Promote warn_dev to WARN for invalid calls in drm_panel.c
This patch is the second of a set of two.

In the kernel docs there is a todo:
Link: https://www.kernel.org/doc/html/v6.8/gpu/todo.html#clean-up-checks-for-already-prepared-enabled-in-panels

The first patch addresses removing the duplicated code. This second
patch addresses promoting instances of dev_warn to WARN as it is an
error in other code for these functions to be called in the invalid
state that triggers the checks.

Signed-off-by: Emilio Mendoza Reyes <emendoz@clemson.edu>